creating a Form directly
