+++ src/main/java/com/training/Mybatiswork0920/Form/UpdateForm.java
@@ -1,0 +1,32 @@
+package com.training.Mybatiswork0920.Form;
+
+public class UpdateForm {
+    private String ex_in;
+    private String category;
+    private int amount;
+
+    public String getEx_in() {
+        return ex_in;
+    }
+
+    public void setEx_in(String ex_in) {
+        this.ex_in = ex_in;
+    }
+
+    public String getCategory() {
+        return category;
+    }
+
+    public void setCategory(String category) {
+        this.category = category;
+    }
+
+    public int getAmount() {
+        return amount;
+    }
+
+    public void setAmount(int amount) {
+        this.amount = amount;
+    }
+
+}
